pyproject.toml
[tool.poetry]
name = "backend-core"
version = "0.1.0"
description = "The best core for the best backend!"
authors = ["adrian.sacristan <adriansacrista1993n@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
psycopg2 = "^2.9.6"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
sqladmin = "^0.14.0"
itsdangerous = "^2.1.2"
pydantic = {version = "1.10.12", extras = ["dotenv", "email"]}
alembic = "^1.11.2"
asyncpg = "^0.28.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
sqlmodel = "^0.0.8"
httpx = "^0.24.1"
asyncio = "^3.4.3"
pytest-asyncio = "^0.21.1"
pytest-mock = "^3.11.1"
pre-commit = "^3.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
